Fix: Implement token value formatting (prefix/suffix) for all input types
This commit restores and properly implements token value formatting that was
broken in the previous refactoring. The formatting system allows dashboard
inputs to manipulate token values with prefix/suffix modifiers.

## Changes

### 1. Added Formatting Properties to InputOptions
- Added `prefix`, `suffix` properties for all input types
- Added `valuePrefix`, `valueSuffix`, `delimiter` for multiselect inputs
- Properties are parsed from DashboardInput.optionsJSON

### 2. Enhanced TokenAdapter
- Exposed formatting properties from InputOptions
- Created `formatTokenValue()` method that handles:
  - Single value formatting: applies prefix/suffix
  - Multiselect formatting: applies valuePrefix/valueSuffix to each value,
    joins with delimiter, then applies outer prefix/suffix

### 3. Updated TokenInputView
- Modified `saveTokenValue()` to apply formatting before saving
- Modified `createBinding()` for multiselect to use array-based formatting
- Added debug logging to track raw vs formatted values
- Change handlers now receive raw values (not formatted)

### 4. Fixed Type Inconsistency
- Changed SearchExecution.resultCount from Int64 to Int
- Updated SearchExecutionAdapter to match
- Fixed CoreDataManager assignment

## Example Usage

For a multiselect input with:
- valuePrefix: "file=\""
- valueSuffix: "\""
- delimiter: " OR "
- Selected values: ["error.log", "access.log"]

Formatted token value: `file="error.log" OR file="access.log"`

## Files Modified
- SplTv/Models/InputChoice.swift
- SplTv/Models/TokenAdapter.swift
- SplTv/Views/TokenInputView.swift
- SplTv/Models/SearchExecutionAdapter.swift
- DashboardKit/CoreData/Entities/SearchExecution+CoreDataProperties.swift
- DashboardKit/Managers/CoreDataManager.swift

+++ d8aTv/Sources/DashboardKit/CoreData/Entities/SearchExecution+CoreDataProperties.swift
@@ -14,7 +14,7 @@ extension SearchExecution {
     @NSManaged public var startTime: Date?
     @NSManaged public var endTime: Date?
     @NSManaged public var status: String?
-    @NSManaged public var resultCount: Int64
+    @NSManaged public var resultCount: Int
     @NSManaged public var errorMessage: String?
     @NSManaged public var dataSource: DataSource?
     @NSManaged public var results: NSSet?
